// src/app/shared/enums/partner-type.enum.ts
/** Felhasználó típusokat definiáló enum. */
export enum PartnerType {
  SUPERADMIN = 'SUPERADMIN',
  PRIVATE_CUSTOMER = 'PRIVATE_CUSTOMER',
  EXPERT_ADMIN = 'EXPERT_ADMIN',
  MAIN_ACCOUNT_OWNER = 'MAIN_ACCOUNT_OWNER',
  CORPORATE_CUSTOMER = 'CORPORATE_CUSTOMER',
  CORPORATE_CUSTOMER_ADMIN = 'CORPORATE_CUSTOMER_ADMIN',
  PUBLIC_INSTITUTION_CUSTOMER = 'PUBLIC_INSTITUTION_CUSTOMER',
  PUBLIC_INSTITUTION_CUSTOMER_ADMIN = 'PUBLIC_INSTITUTION_CUSTOMER_ADMIN',
  COMPANY = 'COMPANY',
}